Update to tests

+++ YesGraph/YesGraphSDK/YesGraphSDKTests/Tests/YesGraphSDKAddressBookTests.m
@@ -235,7 +235,7 @@
          mockDic[@"user_id"] = YSGTestClientID;
          NSArray <YSGContact *> *contacts = parsedResponse[@"entries"];
          totalRec += contacts.count;
-         NSLog(@"Total rec now: %lu", totalRec);
+         NSLog(@"Total rec now: %lu", (unsigned long)totalRec);
          // the post request is fired twice, so we divide by 2
          if ((totalRec/2) == desiredSize && wasInvoked)
          {
@@ -303,7 +303,7 @@
          mockDic[@"user_id"] = YSGTestClientID;
          NSArray <YSGContact *> *contacts = parsedResponse[@"entries"];
          totalRec += contacts.count;
-         NSLog(@"Total rec now: %lu", totalRec);
+         NSLog(@"Total rec now: %lu", (unsigned long)totalRec);
          return YES;
      }
    andStubResponseBlock:^OHHTTPStubsResponse * _Nonnull(NSURLRequest * _Nonnull request)
